memory_sidecar/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
insightface>=0.7.3
onnxruntime>=1.17.0
faiss-cpu>=1.7.4
pyannote.audio>=3.1.1
graphiti-core>=0.3.0
mem0ai>=0.1.0
openai>=1.12.0
numpy>=1.26.0
Pillow>=10.2.0
opencv-python-headless>=4.9.0.80
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
